Consolidate RMS plotting into scripts directory
As per change request on PR #101.
